fix(api,mcp): move ROBOTS_TXT out of worker entry modules so workerd can start
